tracing-core: relax `get_default`'s argument to `FnOnce`
This change increases the flexibility of `get_default`.

It was already the case that `get_default`'s argument was only
invoked at-most once. However, the compiler could not prove this,
and thus required that the argument be `FnMut`. This commit
introduces `UnsafeFnMut`, a type that wraps an `FnOnce` and
provides an unsafe interface for invoking it in cases where the
compiler cannot otherwise prove that it's only invoked at-most
once.